+++ CHANGELOG.md
@@ -14,6 +14,7 @@
    separate state context and used as argument in `PdfEncrypt` methods
 - Added `PdfNames` and moved all known names there from `PdfName`
 - `PdfPageCollection`: Methods creating pages now takes PdfPageSize or default inferred from doc
+- `PdfName`: Added `PdfName operator""_nm(const char*, size_t)`
 
 ## Version 0.10.3
 - Fixed big performance regression introduced in 0.10, see #108